move exceptionThrowingIterator to intermediateRowParsingReader
